CLN move class property in utils
