+++ benchopt/base.py
@@ -3,7 +3,7 @@ from collections import namedtuple
 from abc import ABC, abstractmethod
 
 from .config import RAISE_INSTALL_ERROR
-from .class_property import classproperty
+from .utils.class_property import classproperty
 from .utils.shell_cmd import install_in_conda_env
 from .utils.shell_cmd import _run_shell_in_conda_env
 from .utils.shell_cmd import shell_install_in_conda_env
